reorganize alert component directory
